refactor(ControlledStepFormDialog): useStepFormDialogButtonをControlledStepFormDialogに統合
useStepFormDialogButtonはControlledStepFormDialogからしか利用されていなかったため、
別ファイルに切り出さずに利用元へ直接実装する方針に合わせて統合した。
